cap_flow_orchestrator/requirements.txt
pymongo>=4.7.3
requests>=2.32.3
boto3>=1.35.63
python-dotenv>=1.0.1
